change getObjectById parameter from int to long

+++ src/main/java/cz/muni/fi/pv168/seminar01/beta/ui/model/ShareCarRiderTableModel.java
@@ -62,7 +62,7 @@ public abstract class ShareCarRiderTableModel<T extends HasID> extends AbstractT
         fireTableRowsUpdated(rowIndex, rowIndex);
     }
 
-    public T getObjectById(int id) {
+    public T getObjectById(long id) {
         return data.stream()
                 .filter(object -> object.getId() == id)
                 .findAny()
